Add event listeners for drawing on canvas
Added event listeners to fillGrid() to allow for
background color change of #canvas Div's when mouse
event is clicked and dragged. Added backgroundColor()
in support of this.

Amended removeChildNodes() to clearCanvas() and simplified
the removal of child nodes.

Changed max size on slider to 60 as 100 was making
the slider lag due to large number of div's created.