Set initial value rbgPercent darken/lighten to 60
Setting the initial value to 60 allows both
darken and lighten to start in the middle range
of gray shades. Previous initial value changed
the initial background color to upper and lower
end of the gray shades which looks like an extreme
snap in color on mouseover. Setting the value to
60 gives a much smoother result