Add if statement to colorPicker onchange
To resolve an issue where color picker was
removing the active button state of grid lines.
This commit corrects the issue where grid lines
active state was going out of synch with the
display of grid.

Also added additional if/else to gridlines
active button to remove trying to add a class
when prevBtn === null. This prevents an error
message in the console.